fix: small mods in component
