+++ src/internal/tui/component/model.go
@@ -153,7 +153,7 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 			m.help.ShowAll = !m.help.ShowAll
 
 		case common.ContainsKey(k, m.keys.NavigateLeft.Keys()):
-			if !m.componentPicker.Open && !m.frameworkPicker.Open && !m.detailView.Open {
+			if !m.inOverlay() {
 				if m.focus == 0 {
 					m.focus = maxFocus
 				} else {
@@ -163,7 +163,7 @@ func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 			}
 
 		case common.ContainsKey(k, m.keys.NavigateRight.Keys()):
-			if !m.componentPicker.Open && !m.frameworkPicker.Open && !m.detailView.Open {
+			if !m.inOverlay() {
 				m.focus = (m.focus + 1) % (maxFocus + 1)
 				m.updateKeyBindings()
 			}
@@ -514,6 +514,10 @@ func (m *Model) UpdateDescription(description string) {
 	}
 }
 
+func (m *Model) inOverlay() bool {
+	return m.componentPicker.Open || m.frameworkPicker.Open || m.detailView.Open
+}
+
 func (m *Model) resetWidgets() {
 	if m.selectedFramework.OscalFramework != nil {
 		controlItems := make([]blist.Item, len(m.selectedFramework.Controls))
@@ -589,122 +593,4 @@ func (m *Model) updateSizing(height, width int) {
 
 	m.validationPicker.Height = validationsHeight
 	m.validationPicker.Width = rightWidth
-}
-
-func (m *Model) updateKeyBindings() {
-	m.outOfFocus()
-	m.updateFocusHelpKeys()
-
-	switch m.focus {
-
-	case focusControls:
-		m.controls.KeyMap = common.FocusedListKeyMap()
-		m.controls.SetDelegate(common.NewFocusedDelegate())
-
-	case focusValidations:
-		m.validations.KeyMap = common.FocusedListKeyMap()
-		m.validations.SetDelegate(common.NewFocusedDelegate())
-
-	case focusRemarks:
-		m.remarks.KeyMap = common.FocusedPanelKeyMap()
-		m.remarks.MouseWheelEnabled = true
-		if m.remarksEditor.Focused() {
-			m.remarksEditor.KeyMap = common.FocusedTextAreaKeyMap()
-			m.keys = componentEditKeys
-		} else {
-			m.remarksEditor.KeyMap = common.UnfocusedTextAreaKeyMap()
-			m.keys = componentKeys
-		}
-
-	case focusDescription:
-		m.description.KeyMap = common.FocusedPanelKeyMap()
-		m.description.MouseWheelEnabled = true
-		if m.descriptionEditor.Focused() {
-			m.descriptionEditor.KeyMap = common.FocusedTextAreaKeyMap()
-			m.keys = componentEditKeys
-		} else {
-			m.descriptionEditor.KeyMap = common.UnfocusedTextAreaKeyMap()
-			m.keys = componentKeys
-		}
-
-	}
-}
-
-// func for outOfFocus to run just when focus switches between items
-func (m *Model) outOfFocus() {
-	focusMinusOne := m.focus - 1
-	focusPlusOne := m.focus + 1
-
-	if m.focus == 0 {
-		focusMinusOne = maxFocus
-	}
-	if m.focus == maxFocus {
-		focusPlusOne = 0
-	}
-
-	for _, f := range []focus{focusMinusOne, focusPlusOne} {
-		// Turn off keys for out of focus items
-		switch f {
-		case focusControls:
-			m.controls.KeyMap = common.UnfocusedListKeyMap()
-
-		case focusValidations:
-			m.validations.KeyMap = common.UnfocusedListKeyMap()
-			m.validations.SetDelegate(common.NewUnfocusedDelegate())
-			m.validations.ResetSelected()
-
-		case focusRemarks:
-			m.remarks.KeyMap = common.UnfocusedPanelKeyMap()
-			m.remarks.MouseWheelEnabled = false
-
-		case focusDescription:
-			m.description.KeyMap = common.UnfocusedPanelKeyMap()
-			m.description.MouseWheelEnabled = false
-		}
-	}
-}
-
-func (m *Model) updateFocusHelpKeys() {
-	switch m.focus {
-	case focusComponentSelection:
-		m.help.ShortHelp = shortHelpDialogBox
-		m.help.FullHelpOneLine = fullHelpDialogBoxOneLine
-		m.help.FullHelp = fullHelpDialogBox
-	case focusFrameworkSelection:
-		m.help.ShortHelp = shortHelpDialogBox
-		m.help.FullHelpOneLine = fullHelpDialogBoxOneLine
-		m.help.FullHelp = fullHelpDialogBox
-	case focusControls:
-		m.help.ShortHelp = common.ShortHelpList
-		m.help.FullHelpOneLine = common.FullHelpListOneLine
-		m.help.FullHelp = common.FullHelpList
-	case focusRemarks:
-		if m.remarksEditor.Focused() {
-			m.help.ShortHelp = common.ShortHelpEditing
-			m.help.FullHelpOneLine = common.FullHelpEditingOneLine
-			m.help.FullHelp = common.FullHelpEditing
-		} else {
-			m.help.ShortHelp = shortHelpEditableDialogBox
-			m.help.FullHelpOneLine = fullHelpEditableDialogBoxOneLine
-			m.help.FullHelp = fullHelpEditableDialogBox
-		}
-	case focusDescription:
-		if m.descriptionEditor.Focused() {
-			m.help.ShortHelp = common.ShortHelpEditing
-			m.help.FullHelpOneLine = common.FullHelpEditingOneLine
-			m.help.FullHelp = common.FullHelpEditing
-		} else {
-			m.help.ShortHelp = shortHelpEditableDialogBox
-			m.help.FullHelpOneLine = fullHelpEditableDialogBoxOneLine
-			m.help.FullHelp = fullHelpEditableDialogBox
-		}
-	case focusValidations:
-		m.help.ShortHelp = shortHelpValidations
-		m.help.FullHelpOneLine = fullHelpValidationsOneLine
-		m.help.FullHelp = fullHelpValidations
-	default:
-		m.help.ShortHelp = shortHelpNoFocus
-		m.help.FullHelpOneLine = fullHelpNoFocusOneLine
-		m.help.FullHelp = fullHelpNoFocus
-	}
 }
